Move view modifiers to file
