+++ hackertracker/Views/EventDetailView.swift
@@ -57,20 +57,6 @@ struct EventDetailView: View {
     }
 }
 
-struct RectangleBackground: ViewModifier {
-    func body(content: Content) -> some View {
-        content.padding(10).multilineTextAlignment(.leading)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(Rectangle().fill(Color(UIColor(hex: "#2d2d2D") ?? UIColor.gray))).cornerRadius(5)
-    }
-}
-
-extension View {
-    func rectangleBackground() -> some View {
-        modifier(RectangleBackground())
-    }
-}
-
 struct EventDetailView_Previews: PreviewProvider {
     static var previews: some View {
         NavigationView {
